Restructure source code files and add dependencies
The source directories are now matched up so that they follow the Maven
directory guidelines. Additionally, the local Spigot library that the
project was using so far was substituted for a remote Bukkit library.
This was mainly done as to refrain from using local .jar files as
dependencies, since binary files should not exist in VCS.